Cleanup all one-off rewards hooks, update forms to show net apys
