--- apps/hyperdrive-trading/src/ui/rewards/useAeroRate.ts
@@ -1,594 +1,0 @@
-import { fixed, FixedPoint, parseFixed } from "@delvtech/fixed-point-wasm";
-import { HyperdriveConfig } from "@delvtech/hyperdrive-appconfig";
-import { useTokenFiatPrice } from "src/ui/token/hooks/useTokenFiatPrice";
-import { useReadContract } from "wagmi";
-
-/**
- * @deprecated This now comes from the rewards resolver directly
- */
-export function useAeroRate({
-  hyperdrive,
-  enabled,
-}: {
-  hyperdrive: HyperdriveConfig;
-  enabled: boolean;
-}): {
-  aeroRate: FixedPoint | undefined;
-} {
-  // Amount of AERO rewarded per second (~1.9 AERO per second)
-  const { data } = useReadContract({
-    address: "0x4F09bAb2f0E15e2A078A227FE1537665F55b8360",
-    chainId: hyperdrive.chainId,
-    functionName: "rewardRate",
-    abi: gaugeAbi,
-    query: {
-      enabled,
-    },
-  });
-  const fixedPointRate = fixed(data ?? 0n);
-
-  // Multiply this by the number of seconds in a year
-  const secondsPerYear = 31_536_000;
-  const fixedSecondsPerYear = parseFixed(secondsPerYear);
-  const aeroRewardsPerYear = fixedPointRate.mul(fixedSecondsPerYear);
-
-  // price per AERO token
-  const { fiatPrice } = useTokenFiatPrice({
-    tokenAddress: "0x940181a94A35A4569E4529A3CDfB74e38FD98631",
-    chainId: hyperdrive.chainId,
-  });
-
-  const dollarAmountRewardedPerYear = aeroRewardsPerYear.mul(
-    fixed(fiatPrice ?? 0n),
-  );
-
-  const { data: totalSupply } = useReadContract({
-    address: "0x4F09bAb2f0E15e2A078A227FE1537665F55b8360",
-    chainId: hyperdrive.chainId,
-    functionName: "totalSupply",
-    abi: gaugeAbi,
-    query: {
-      enabled,
-    },
-  });
-
-  const fixedTotalSupply = fixed(totalSupply ?? 0n);
-
-  const { fiatPrice: fiatPriceOfLPToken } = useTokenFiatPrice({
-    tokenAddress: "0x6cDcb1C4A4D1C3C6d054b27AC5B77e89eAFb971d",
-    chainId: hyperdrive.chainId,
-  });
-
-  const dollarValueOfPool = fixed(fiatPriceOfLPToken ?? 0n).mul(
-    fixedTotalSupply,
-  );
-
-  let apr = fixed(0n);
-  if (!dollarValueOfPool.eq(0n)) {
-    apr = dollarAmountRewardedPerYear
-      .div(dollarValueOfPool)
-      .mul(parseFixed(100n));
-  }
-  return { aeroRate: enabled ? fixed(apr) : undefined };
-}
-
-export const gaugeAbi = [
-  {
-    inputs: [],
-    name: "NotAlive",
-    type: "error",
-  },
-  {
-    inputs: [],
-    name: "NotAuthorized",
-    type: "error",
-  },
-  {
-    inputs: [],
-    name: "NotTeam",
-    type: "error",
-  },
-  {
-    inputs: [],
-    name: "NotVoter",
-    type: "error",
-  },
-  {
-    inputs: [],
-    name: "RewardRateTooHigh",
-    type: "error",
-  },
-  {
-    inputs: [],
-    name: "ZeroAmount",
-    type: "error",
-  },
-  {
-    inputs: [],
-    name: "ZeroRewardRate",
-    type: "error",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: true,
-        internalType: "address",
-        name: "from",
-        type: "address",
-      },
-      {
-        indexed: false,
-        internalType: "uint256",
-        name: "claimed0",
-        type: "uint256",
-      },
-      {
-        indexed: false,
-        internalType: "uint256",
-        name: "claimed1",
-        type: "uint256",
-      },
-    ],
-    name: "ClaimFees",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: true,
-        internalType: "address",
-        name: "from",
-        type: "address",
-      },
-      {
-        indexed: false,
-        internalType: "uint256",
-        name: "amount",
-        type: "uint256",
-      },
-    ],
-    name: "ClaimRewards",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: true,
-        internalType: "address",
-        name: "from",
-        type: "address",
-      },
-      {
-        indexed: true,
-        internalType: "address",
-        name: "to",
-        type: "address",
-      },
-      {
-        indexed: false,
-        internalType: "uint256",
-        name: "amount",
-        type: "uint256",
-      },
-    ],
-    name: "Deposit",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: true,
-        internalType: "address",
-        name: "from",
-        type: "address",
-      },
-      {
-        indexed: false,
-        internalType: "uint256",
-        name: "amount",
-        type: "uint256",
-      },
-    ],
-    name: "NotifyReward",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: true,
-        internalType: "address",
-        name: "from",
-        type: "address",
-      },
-      {
-        indexed: false,
-        internalType: "uint256",
-        name: "amount",
-        type: "uint256",
-      },
-    ],
-    name: "Withdraw",
-    type: "event",
-  },
-  {
-    inputs: [
-      {
-        internalType: "address",
-        name: "",
-        type: "address",
-      },
-    ],
-    name: "balanceOf",
-    outputs: [
-      {
-        internalType: "uint256",
-        name: "",
-        type: "uint256",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        internalType: "uint256",
-        name: "_amount",
-        type: "uint256",
-      },
-      {
-        internalType: "address",
-        name: "_recipient",
-        type: "address",
-      },
-    ],
-    name: "deposit",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        internalType: "uint256",
-        name: "_amount",
-        type: "uint256",
-      },
-    ],
-    name: "deposit",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        internalType: "address",
-        name: "_account",
-        type: "address",
-      },
-    ],
-    name: "earned",
-    outputs: [
-      {
-        internalType: "uint256",
-        name: "_earned",
-        type: "uint256",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "fees0",
-    outputs: [
-      {
-        internalType: "uint256",
-        name: "",
-        type: "uint256",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "fees1",
-    outputs: [
-      {
-        internalType: "uint256",
-        name: "",
-        type: "uint256",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "feesVotingReward",
-    outputs: [
-      {
-        internalType: "address",
-        name: "",
-        type: "address",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-
-  {
-    inputs: [],
-    name: "isPool",
-    outputs: [
-      {
-        internalType: "bool",
-        name: "",
-        type: "bool",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "lastTimeRewardApplicable",
-    outputs: [
-      {
-        internalType: "uint256",
-        name: "_time",
-        type: "uint256",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "lastUpdateTime",
-    outputs: [
-      {
-        internalType: "uint256",
-        name: "",
-        type: "uint256",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "left",
-    outputs: [
-      {
-        internalType: "uint256",
-        name: "_left",
-        type: "uint256",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        internalType: "uint256",
-        name: "amount",
-        type: "uint256",
-      },
-    ],
-    name: "notifyRewardAmount",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        internalType: "uint256",
-        name: "amount",
-        type: "uint256",
-      },
-    ],
-    name: "notifyRewardWithoutClaim",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "periodFinish",
-    outputs: [
-      {
-        internalType: "uint256",
-        name: "",
-        type: "uint256",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "rewardPerToken",
-    outputs: [
-      {
-        internalType: "uint256",
-        name: "_rewardPerToken",
-        type: "uint256",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "rewardPerTokenStored",
-    outputs: [
-      {
-        internalType: "uint256",
-        name: "",
-        type: "uint256",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "rewardRate",
-    outputs: [
-      {
-        internalType: "uint256",
-        name: "",
-        type: "uint256",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        internalType: "uint256",
-        name: "",
-        type: "uint256",
-      },
-    ],
-    name: "rewardRateByEpoch",
-    outputs: [
-      {
-        internalType: "uint256",
-        name: "",
-        type: "uint256",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "rewardToken",
-    outputs: [
-      {
-        internalType: "address",
-        name: "",
-        type: "address",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        internalType: "address",
-        name: "",
-        type: "address",
-      },
-    ],
-    name: "rewards",
-    outputs: [
-      {
-        internalType: "uint256",
-        name: "",
-        type: "uint256",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "stakingToken",
-    outputs: [
-      {
-        internalType: "address",
-        name: "",
-        type: "address",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "totalSupply",
-    outputs: [
-      {
-        internalType: "uint256",
-        name: "",
-        type: "uint256",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        internalType: "address",
-        name: "",
-        type: "address",
-      },
-    ],
-    name: "userRewardPerTokenPaid",
-    outputs: [
-      {
-        internalType: "uint256",
-        name: "",
-        type: "uint256",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "ve",
-    outputs: [
-      {
-        internalType: "address",
-        name: "",
-        type: "address",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "voter",
-    outputs: [
-      {
-        internalType: "address",
-        name: "",
-        type: "address",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        internalType: "uint256",
-        name: "_amount",
-        type: "uint256",
-      },
-    ],
-    name: "withdraw",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-] as const;
-
-export type GaugeAbi = typeof gaugeAbi;
