+++ apps/hyperdrive-trading/src/ui/rewards/useRewards.ts
@@ -1,179 +1,42 @@
 import {
-  AERO_ICON_URL,
+  AnyReward,
+  appConfig,
   HyperdriveConfig,
-  WELL_ICON_URL,
+  rewardFunctions,
 } from "@delvtech/hyperdrive-appconfig";
-import {
-  useMorphoRate,
-  useMorphoVaultRewards,
-  vaultAddresses,
-} from "src/ui/rewards/useMorphoRate";
-import { Address } from "viem";
-import { base, mainnet } from "viem/chains";
-import { useAeroRate } from "./useAeroRate";
+import { useQuery } from "@tanstack/react-query";
+import { getPublicClient } from "@wagmi/core";
+import { makeQueryKey } from "src/base/makeQueryKey";
+import { wagmiConfig } from "src/network/wagmiClient";
+import { PublicClient } from "viem";
 
-// TODO @cashd: Move to AppConfig
-// https://github.com/delvtech/hyperdrive-frontend/issues/1341
-const eligibleMarketsForMorphoRewards: Record<number, Address[]> = {
-  [mainnet.id]: [
-    // 182d Morpho wstETH/USDC
-    "0xc8D47DE20F7053Cc02504600596A647A482Bbc46",
-    // 182d Morpho Blue wstETH/USDA
-    "0x7548c4F665402BAb3a4298B88527824B7b18Fe27",
-  ],
-  [base.id]: [
-    // 182d Morpho cbETH/USDC
-    // "0xFcdaF9A4A731C24ed2E1BFd6FA918d9CF7F50137",
-    // 182d Moonwell Flagship ETH
-    // "0xceD9F810098f8329472AEFbaa1112534E96A5c7b",
-    // 182d Moonwell Flagship USDC
-    // "0xD9b66D9a819B36ECEfC26B043eF3B422d5A6123a",
-    // 182d Moonwell Flagship EURC
-    // "0xdd8E1B14A04cbdD98dfcAF3F0Db84A80Bfb8FC25",
-  ],
-};
+export function useRewards(hyperdriveConfig: HyperdriveConfig): {
+  rewards: AnyReward[] | undefined;
+  status: "error" | "success" | "loading";
+} {
+  const rewardsFn =
+    appConfig.yieldSources[hyperdriveConfig.yieldSource].rewardsFn;
 
-export const eligibleMarketsForMorphoVaultRewards: Record<number, Address[]> = {
-  [base.id]: [
-    // 182d Moonwell Flagship ETH
-    "0xceD9F810098f8329472AEFbaa1112534E96A5c7b",
-    // 182d Moonwell Flagship USDC
-    "0xD9b66D9a819B36ECEfC26B043eF3B422d5A6123a",
-    // 182d Moonwell Flagship EURC
-    "0xdd8E1B14A04cbdD98dfcAF3F0Db84A80Bfb8FC25",
-  ],
-};
+  const queryEnabled = !!rewardsFn;
 
-export const eligibleForAeroRewards: Record<number, Address[]> = {
-  [base.id]: [
-    // 182d USDC/AERO market
-    // TODO: Update to correct market when deployed
-    "0x9bAdB6A21FbA04EE94fde3E85F7d170E90394c89",
-  ],
-};
-
-type RewardType =
-  | "MorphoFlatRate"
-  | "MorphoVault"
-  | "MorphoVaultAllocation"
-  | "Aero";
-
-type Reward = {
-  id: RewardType;
-  name: string;
-  amount: string;
-  iconUrl?: string;
-};
-
-/**
- * @deprecated use useAppConfigRewards instead
- */
-
-export function useRewards(hyperdrive: HyperdriveConfig): Reward[] | undefined {
-  const { morphoRate } = useMorphoRate({
-    chainId: hyperdrive.chainId,
-    hyperdriveAddress: hyperdrive.address,
-    enabled:
-      eligibleMarketsForMorphoRewards[hyperdrive.chainId]?.includes(
-        hyperdrive.address,
-      ) ?? false,
-  });
-
-  const { morphoVaultData } = useMorphoVaultRewards({
-    hyperdrive,
-    enabled:
-      eligibleMarketsForMorphoVaultRewards[base.id]?.includes(
-        hyperdrive.address,
-      ) ?? false,
-  });
-
-  const { aeroRate } = useAeroRate({
-    enabled:
-      eligibleForAeroRewards[base.id]?.includes(hyperdrive.address) ?? false,
-    hyperdrive,
-  });
-
-  const rewards = [];
-
-  if (
-    eligibleForAeroRewards[hyperdrive.chainId]?.includes(hyperdrive.address)
-  ) {
-    const aeroReward: Reward = {
-      id: "Aero",
-      name: "AERO",
-      iconUrl: AERO_ICON_URL,
-      amount: aeroRate
-        ? aeroRate.format({
-            decimals: 2,
-          })
-        : "0",
-    };
-    rewards.push(aeroReward);
-  }
-
-  // Add any morpho rewards for this market
-  if (
-    eligibleMarketsForMorphoRewards[hyperdrive.chainId]?.includes(
-      hyperdrive.address,
-    )
-  ) {
-    const morphoReward: Reward = {
-      id: "MorphoFlatRate",
-      name: "MORPHO",
-      amount: morphoRate
-        ? morphoRate.format({
-            decimals: 2,
-          })
-        : "0",
-    };
-    rewards.push(morphoReward);
-  }
-  if (
-    eligibleMarketsForMorphoVaultRewards[base.id]?.includes(hyperdrive.address)
-  ) {
-    const vaultReward: Reward = {
-      id: "MorphoVault",
-      name: morphoVaultData?.vault?.state.rewards[0].asset.name ?? "WELL",
-      iconUrl: WELL_ICON_URL,
-      amount: morphoVaultData?.reward?.supplyApr
-        ? `${(morphoVaultData.reward.supplyApr * 100).toFixed(2)}%`
-        : "0%",
-    };
-
-    if (vaultAddresses[hyperdrive.address].allocation) {
-      let vaultAllocationRewardTotal: number = 0;
-
-      const totalAssetsUsd = morphoVaultData.vault?.state.totalAssetsUsd;
-
-      morphoVaultData.vault?.state.allocation?.forEach((allocation) => {
-        const rewardsMarket = allocation.market.state.rewards.find(
-          (reward) =>
-            reward.asset.address ===
-            vaultAddresses[hyperdrive.address].allocation?.assets[0].address,
-        );
-
-        if (rewardsMarket) {
-          const marketRewardsEarned =
-            allocation.supplyAssetsUsd * rewardsMarket.supplyApr;
-          vaultAllocationRewardTotal += marketRewardsEarned;
+  const { data: rewards, status } = useQuery({
+    queryKey: makeQueryKey("rewards", {
+      hyperdriveAddress: hyperdriveConfig.address,
+      chainId: hyperdriveConfig.chainId,
+    }),
+    enabled: queryEnabled,
+    queryFn: queryEnabled
+      ? async () => {
+          const publicClient = getPublicClient(wagmiConfig as any, {
+            chainId: hyperdriveConfig.chainId,
+          }) as PublicClient;
+          return rewardFunctions[rewardsFn](publicClient);
         }
-      });
+      : undefined,
+  });
 
-      const vaultAllocationReward: Reward = {
-        id: "MorphoVaultAllocation",
-        name:
-          vaultAddresses[hyperdrive.address].allocation?.assets[0].name ?? "",
-        iconUrl:
-          vaultAddresses[hyperdrive.address].allocation?.assets[0].assetIcon,
-        // Vault allocation rewards are calculated by taking the total assets in the vault (in USD), and dividing it by the rewards earned for the vault in each market it has allocated to.
-        amount: totalAssetsUsd
-          ? `${((vaultAllocationRewardTotal / totalAssetsUsd) * 100).toFixed(2)}%`
-          : "0%",
-      };
-      rewards.push(vaultAllocationReward);
-    }
-    rewards.push(vaultReward);
-  }
-
-  return rewards;
+  return {
+    rewards,
+    status,
+  };
 }
